Refactor: 점진적 마이그레이션 (글로벌 모달l)
